Deler LetterRoutes i to for å separere autobrev frå redigerbare, det gjer i neste runde at vi kan forenkle meir innover
