+++ pensjon-brevbaker/src/main/kotlin/no/nav/pensjon/brev/routing/LetterRoutes.kt
@@ -6,10 +6,8 @@ import io.ktor.server.routing.*
 import io.ktor.util.AttributeKey
 import no.nav.pensjon.brev.api.TemplateResource
 import no.nav.pensjon.brev.api.model.BestillBrevRequest
-import no.nav.pensjon.brev.api.model.BestillRedigertBrevRequest
 import no.nav.pensjon.brev.api.model.maler.Brevkode
 import no.nav.pensjon.brev.template.AutobrevTemplate
-import no.nav.pensjon.brev.template.RedigerbarTemplate
 
 private val BREV_KODE = AttributeKey<String>("CallStartTime")
 fun <T : Brevkode<T>> RoutingContext.installBrevkodeInCallContext(kode: Brevkode<T>) {
@@ -17,9 +15,8 @@ fun <T : Brevkode<T>> RoutingContext.installBrevkodeInCallContext(kode: Brevkode
 }
 fun ApplicationCall.useBrevkodeFromCallContext(): String? = attributes.getOrNull(BREV_KODE)
 
-fun Route.letterRoutes(
+fun Route.autobrevRoutes(
     autobrev: TemplateResource<Brevkode.Automatisk, AutobrevTemplate<*>>,
-    redigerbareBrev: TemplateResource<Brevkode.Redigerbart, RedigerbarTemplate<*>>,
 ) {
     route("/${autobrev.name}") {
         post<BestillBrevRequest<Brevkode.Automatisk>>("/pdf") { brevbestilling ->
@@ -35,22 +32,6 @@ fun Route.letterRoutes(
 
         post<BestillBrevRequest<Brevkode.Automatisk>>("/json") {
             call.respond(autobrev.renderJSON(it))
-        }
-    }
-    route("/${redigerbareBrev.name}") {
-        post<BestillBrevRequest<Brevkode.Redigerbart>>("/markup") { brevbestilling ->
-            val markup = redigerbareBrev.renderLetterMarkup(brevbestilling)
-            call.respond(markup)
-        }
-        post<BestillRedigertBrevRequest<Brevkode.Redigerbart>>("/pdf") { brevbestilling ->
-            installBrevkodeInCallContext(brevbestilling.kode)
-            call.respond(redigerbareBrev.renderPDF(brevbestilling))
-            redigerbareBrev.countLetter(brevbestilling.kode)
-        }
-
-        post<BestillRedigertBrevRequest<Brevkode.Redigerbart>>("/html") { brevbestilling ->
-            call.respond(redigerbareBrev.renderHTML(brevbestilling))
-            redigerbareBrev.countLetter(brevbestilling.kode)
         }
     }
 }
