+++ pensjon-brevbaker/src/main/kotlin/no/nav/pensjon/brev/routing/RedigerbarRoutes.kt
@@ -1,0 +1,33 @@
+package no.nav.pensjon.brev.routing
+
+import io.ktor.server.response.respond
+import io.ktor.server.routing.Route
+import io.ktor.server.routing.post
+import io.ktor.server.routing.route
+import no.nav.pensjon.brev.api.TemplateResource
+import no.nav.pensjon.brev.api.model.BestillBrevRequest
+import no.nav.pensjon.brev.api.model.BestillRedigertBrevRequest
+import no.nav.pensjon.brev.api.model.maler.Brevkode
+import no.nav.pensjon.brev.template.RedigerbarTemplate
+
+
+fun Route.redigerbarRoutes(
+    redigerbareBrev: TemplateResource<Brevkode.Redigerbart, RedigerbarTemplate<*>>,
+) {
+    route("/${redigerbareBrev.name}") {
+        post<BestillBrevRequest<Brevkode.Redigerbart>>("/markup") { brevbestilling ->
+            val markup = redigerbareBrev.renderLetterMarkup(brevbestilling)
+            call.respond(markup)
+        }
+        post<BestillRedigertBrevRequest<Brevkode.Redigerbart>>("/pdf") { brevbestilling ->
+            installBrevkodeInCallContext(brevbestilling.kode)
+            call.respond(redigerbareBrev.renderPDF(brevbestilling))
+            redigerbareBrev.countLetter(brevbestilling.kode)
+        }
+
+        post<BestillRedigertBrevRequest<Brevkode.Redigerbart>>("/html") { brevbestilling ->
+            call.respond(redigerbareBrev.renderHTML(brevbestilling))
+            redigerbareBrev.countLetter(brevbestilling.kode)
+        }
+    }
+}
